requirements.txt
panda
numpy
seaborn
matplotlib
ipykernel
scikit-learn
catboost
xgboost
dill
#-e .